feat(modal): remove Modal.Form
