+++ src/modal/index.scss
